requirements.txt
aiosqlite==0.16.1
autopep8==1.5.5
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
fastapi==0.63.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.0
iso8601==0.1.14
passlib==1.7.4
priority==1.3.0
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7.3
PyJWT==2.0.1
PyPika==0.44.1
python-multipart==0.0.5
pytz==2020.5
six==1.15.0
starlette==0.13.6
toml==0.10.2
tortoise-orm==0.16.21
typing-extensions==3.7.4.3
uvicorn==0.13.4
wsproto==1.0.0
